+++ docs/changelog.rst
@@ -5,6 +5,10 @@ Changelog
 
 Dev version
 -----------
+*PR #306*
+
+* Add Q-learning agent in :class:`rlberry.agents.QLAgent` and SARSA agent in :class:`rlberry.agents.SARSAAgent`.
+
 *PR #298*
 
 * Move old scripts (jax agents, attention networks, old examples...) that we won't maintain from the main branch to an archive branch.